Increase the MTU of the transmitted frames from 1496 to 1500 which is the maximum non-jumbo MTU.
On investigation when the IEEE 802.3ac specification added the VLAN tag, the maximum frame length was increased by 4 octets to allow for the encapsulated VLAN tag.

The change means the ports on the switch under test will now see the maximum MTU, after the injection switch has removed the VLAN tag.

Did notice a quirk with Windows where needed to enable jumbo frames to allow for the additional 4 octets in the frames transmitted by the PC. Without enabling jumbo frames in the network adapter settings the send function was returning an error. Haven't traced the PCAP code to determine if this is a bug or expected behavior.